+++ bol/release_notes.md
@@ -1,3 +1,23 @@
+1.0.5 (14-May-2025):
+
+-  Resolve the problem with stock updating.
+
+1.0.4 (21-April-2025):
+
+-  Fixed inventory import to use FBB warehouse instead of default.
+
+1.0.3 (18-March-2025):
+
+-  Add a customerComments when importing BOL returns.
+
+1.0.2 (12-March-2025):
+
+- Fixed Cancel order issue.
+
+1.0.1 (21-Feb-2025):
+
+- Fix bug in a handled return process.
+
 1.0 (4-Dec-2024):
 
 - Initial Release for Odoo v18.